removing duplicate installations and installations from test account
